new proj command
